requirements.txt
azure-iot-device
python-dotenv
azure-eventhub